First commit
Project is quite messy at this point, but backup is needed, and should be
online. The purpose of this project is to record snapshots of position and
time with sufficient precision for a part of our masters thesis. At the
moment there are buttons, an upopulated table, and great plans. Maybe
someday history will look neat and this text will not exist any more.