Fixes relative possitioning issue of table
The table would go behind the trigger button if it got too long. Now it
will not show behind the button, but still be scrollable in the way
intended.

Just occurred to me: Row fonts should perhaps be a bit larger, and the
rows should have a little bit more spacing between them. (In addittion to
the other features to distinguish them that were tought of.) As of now
there are more important things to fix, so this might be future simons
problem.